requirements.txt
influxdb-client ~= 1.36
lowatt-grdf ~= 2.1
python-dotenv ~= 0.21
schedule ~= 1.1
sentry-sdk ~= 1.15
python-snap7 ~= 1.3
